% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv-path.R
\name{available}
\alias{available}
\title{Check whether uv is available}
\usage{
available()
}
\value{
\code{TRUE} if \code{uv} can be found, \code{FALSE} otherwise.
}
\description{
Tests whether the \code{uv} executable can be located on the system.
}
\examples{
available()

}
\seealso{
\code{\link[=uv_path]{uv_path()}}, \code{\link[=install_uv]{install_uv()}}
}
