% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv-path.R
\name{check_available}
\alias{check_available}
\title{Require uv to be available}
\usage{
check_available()
}
\value{
Invisibly returns \code{NULL} if \code{uv} is available.
}
\description{
Aborts with an informative message if the \code{uv} executable cannot be found.
}
\examples{
\dontrun{check_available()}

}
\keyword{internal}
