% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lock.R
\name{lock_path}
\alias{lock_path}
\title{Get the path to the lockfile}
\usage{
lock_path(path = ".")
}
\arguments{
\item{path}{Directory inside a uv project.}
}
\value{
A character string giving the path to \code{uv.lock}.
}
\description{
Returns the path to the \code{uv.lock} file for the project.
}
