% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyproject.R
\name{pyproject_path}
\alias{pyproject_path}
\title{Locate the pyproject.toml file for a uv project}
\usage{
pyproject_path(path = ".")
}
\arguments{
\item{path}{Directory inside a uv project.}
}
\value{
A character string giving the path to \code{pyproject.toml}.
}
\description{
Returns the path to the \code{pyproject.toml} file for the uv project containing
\code{path}. Errors if no uv project is found.
}
