% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{python}
\alias{python}
\title{Get Python interpreter used by uv}
\usage{
python(path = ".")
}
\arguments{
\item{path}{Directory inside a uv project.}
}
\value{
A character string giving the path to the Python interpreter.
}
\description{
Returns the path to the Python interpreter selected by uv for a project.
}
\examples{
\dontshow{if (uv::available()) withAutoprint(\{ # examplesIf}
withr::local_tempdir() |>
  withr::with_dir({
    init()
    python()
  })
\dontshow{\}) # examplesIf}
}
