% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{python_version}
\alias{python_version}
\title{Get Python version used by uv}
\usage{
python_version(path = ".")
}
\arguments{
\item{path}{Directory inside a uv project.}
}
\value{
A character string giving the Python version.
}
\description{
Returns the version of the Python interpreter selected by uv for a project.
}
\examples{
\dontshow{if (uv::available()) withAutoprint(\{ # examplesIf}
with_temp_project({
    init()
    python_version()
  })
\dontshow{\}) # examplesIf}
}
