% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugggest-edit-file.R
\name{suggest_edit_file}
\alias{suggest_edit_file}
\title{Suggest editing a file}
\usage{
suggest_edit_file(path)
}
\arguments{
\item{path}{Path to a file.}
}
\description{
Prints a clickable hint to edit a file if running interactively.
}
\keyword{internal}
