% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{uses_project}
\alias{uses_project}
\title{Check whether a uv project is present}
\usage{
uses_project(path = ".")
}
\arguments{
\item{path}{Directory to start searching from. Defaults to the current
working directory.}
}
\value{
\code{TRUE} if a uv project is detected, \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if a uv project can be detected starting from the given
directory. A uv project is defined by the presence of both
\code{pyproject.toml} and \code{uv.lock}.
}
\examples{
\dontrun{uses_project()}

}
