% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venv.R
\name{venv}
\alias{venv}
\title{Create or update a Python virtual environment}
\usage{
venv(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical; whether to suppress uv output.}
}
\description{
Create or update a Python virtual environment
}
