% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{with_project_root}
\alias{with_project_root}
\title{Run code with the working directory set to the uv project root}
\usage{
with_project_root(path = ".", code)
}
\arguments{
\item{path}{Directory used to locate the uv project.}

\item{code}{Code to evaluate with the working directory set to the project
root.}
}
\value{
Returns the result of evaluating \code{code}.
}
\description{
Evaluates code with the working directory temporarily set to the root of the
uv project containing the given path.
}
\examples{
\dontrun{
with_project_root({
  uv("lock")
})
}

}
