% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{with_temp_project}
\alias{with_temp_project}
\title{Run code inside a temporary uv project workspace}
\usage{
with_temp_project(code)
}
\arguments{
\item{code}{Code to evaluate inside the temporary project workspace.}
}
\value{
Returns the result of evaluating \code{code}.
}
\description{
Creates a temporary directory, sets it as the working directory and temporary
directory, evaluates code, then deletes the directory.
}
\examples{
with_temp_project({
  cat("Working in:", getwd(), "\n")
})

}
